+++ R/globals.R
@@ -2,4 +2,4 @@ utils::globalVariables(c("itemId", "value", "metric", "data", "name",
                          'dataAnomalyDetected', 'dataLowerBound',
                          'dataUpperBound', 'day', 'eventnm', 'qickview',
                          'columnId', 'filters', 'filtername', 'mfinalname',
-                         'segmentIds', 'metrics'))
+                         'pred_items', 'pred_list', 'segmentIds', 'metrics'))